# services/banking-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing accounts, balances, and banking operations.",
        "title": "Banking Service API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/exchange/calculate": {
            "get": {
                "description": "Converts a given amount from one currency to another using current exchange rates.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exchange"
                ],
                "summary": "Calculate currency conversion",
                "parameters": [
                    {
                        "minimum": 0,
                        "type": "number",
                        "description": "Amount to convert",
                        "name": "amount",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Source currency code (e.g. USD, EUR)",
                        "name": "from_currency",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Target currency code (e.g. RSD, EUR)",
                        "name": "to_currency",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ConvertResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
        "/exchange/rates": {
            "get": {
                "description": "Returns the latest available currency exchange rates.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exchange"
                ],
                "summary": "Get exchange rates",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ExchangeRatesResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ConvertResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "from_currency": {
                    "type": "string"
                },
                "to_currency": {
                    "type": "string"
                },
                "total": {
                    "type": "number"
                }
            }
        },
...
        "dto.ExchangeRatesResponse": {
            "type": "object",
            "properties": {
                "base_currency": {
                    "type": "string"
                },
                "next_update_at": {
                    "type": "string"
                },
                "rates": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.RateItem"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "dto.RateItem": {
            "type": "object",
            "properties": {
                "buy_rate": {
                    "type": "number"
                },
                "currency": {
                    "type": "string"
                },
                "middle_rate": {
                    "type": "number"
                },
                "sell_rate": {
                    "type": "number"
                }
            }
        },
...
        "errors.AppError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Authorization header using the Bearer scheme.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}